+++ src/components/InstanceBrowseFilters/index.js
@@ -1,0 +1,2 @@
+export { default } from './InstanceFilterBrowse';
+export { default as instanceFilterBrowseRenderer } from './instanceFilterBrowseRenderer';